% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_sixel_support.R
\name{has_sixel_support}
\alias{has_sixel_support}
\title{Determine if terminal supports Sixel}
\usage{
has_sixel_support(warn = FALSE, throw = FALSE)
}
\arguments{
\item{warn}{show warnings when the protocol is not supported.}

\item{throw}{throw an error when we are running in a kitty terminal.}
}
\value{
Returns \code{TRUE} if the current terminal supports the Sixel and
\code{FALSE} otherwise. Will also return \code{FALSE} when not running in a
terminal.
}
\description{
Determine if terminal supports Sixel
}
\examples{
if (has_sixel_support()) {
  cat("Yeeh, your terminal supports the sixel!")
}

}
