% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_tgp_support.R
\name{has_tgp_support}
\alias{has_tgp_support}
\title{Determine if terminal supports Terminal Graphics Protocol}
\usage{
has_tgp_support(warn = FALSE, throw = FALSE)
}
\arguments{
\item{warn}{show warnings when the protocol is not supported.}

\item{throw}{throw an error when we are running in a kitty terminal.}
}
\value{
Returns \code{TRUE} if the current terminal supports the Terminal Graphics
Protocol and \code{FALSE} otherwise. Will also return \code{FALSE} when not
running in a terminal.
}
\description{
Determine if terminal supports Terminal Graphics Protocol
}
\examples{
if (has_tgp_support()) {
  cat("Yeeh, your terminal supports the terminal graphics protocol!")
}

}
