% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_kitty.R
\name{is_kitty}
\alias{is_kitty}
\title{Determine if we are running in a kitty terminal}
\usage{
is_kitty()
}
\value{
Returns \code{TRUE} if R is running in a kitty terminal
and \code{FALSE} otherwise.
}
\description{
Determine if we are running in a kitty terminal
}
\examples{
if (is_kitty()) {
  cat("Yeeh, you are running kitty!")
}

}
