% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_terminal.R
\name{is_kitty}
\alias{is_kitty}
\alias{is_ghostty}
\alias{is_nvim}
\alias{is_vim}
\alias{is_tmux}
\title{Determine the type of terminal}
\usage{
is_kitty()

is_ghostty()

is_nvim()

is_vim()

is_tmux()
}
\value{
Returns \code{TRUE} if R is running in a given terminal
and \code{FALSE} otherwise.
}
\description{
Determine the type of terminal
}
\details{
These functions try to detect if R is running in a given terminal.
}
\examples{
if (is_kitty()) {
  cat("Yeeh, you are running kitty!")
}
if (is_ghostty()) {
  cat("Yeeh, you are running ghostty!")
}

}
