% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png2sixel.R
\name{png2sixel}
\alias{png2sixel}
\title{Dump a PNG image to the terminal in Sixel format}
\usage{
png2sixel(filename)
}
\arguments{
\item{filename}{filename of the PNG image}
}
\value{
Called for it's side effect of writing the image to the terminal (standard
out). Returns \code{NULL} invisibly.
}
\description{
Dump a PNG image to the terminal in Sixel format
}
