% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png2terminal.R
\name{png2terminal}
\alias{png2terminal}
\title{Dump a PNG image to the terminal}
\usage{
png2terminal(filename, method = c("auto", "tgp", "sixel"))
}
\arguments{
\item{filename}{filename of the PNG image}

\item{method}{method with which the graphical output is written to the
terminal. In case of 'auto', it is first checked if the Terminal Graphics
Protocol is supported and if so, this is used. Otherwise, Sixel is used.}
}
\value{
Called for it's side effect of writing the image to the terminal (standard
out). Returns \code{NULL} invisibly.
}
\description{
Dump a PNG image to the terminal
}
\seealso{
See \code{\link{png2tgp}} for output in Terminal Graphics Protocol.  See
\code{\link{png2sixel}} for output in Sixel format.
}
