% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster2sixel.R
\name{raster2sixel}
\alias{raster2sixel}
\title{Dump an image raster to the terminal in Sixel format}
\usage{
raster2sixel(raster)
}
\arguments{
\item{raster}{the image 'raster' e.g. the output of \code{\link{as.raster}}.}
}
\value{
Called for it's side effect of writing the image to the terminal (standard
out). Returns \code{NULL} invisibly.
}
\description{
Dump an image raster to the terminal in Sixel format
}
\details{
Sixel is a bitmap format supported by some terminals. See, for example, 
\url{https://en.wikipedia.org/wiki/Sixel}.
}
\seealso{
See \code{\link{raster2tgp}} for output using Terminal Graphics Protocol.
}
