% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster2terminal.R
\name{raster2terminal}
\alias{raster2terminal}
\title{Dump an image raster to the terminal}
\usage{
raster2terminal(raster, method = c("auto", "tgp", "sixel"), ...)
}
\arguments{
\item{raster}{the image 'raster' e.g. the output of \code{\link{as.raster}}.}

\item{method}{method with which the graphical output is written to the
terminal. In case of 'auto', it is first checked if the Terminal Graphics
Protocol is supported and if so, this is used. Otherwise, Sixel is used.}

\item{...}{passed on to the underlying method such as
\code{\link{raster2tgp}} and \code{\link{raster2sixel}}.}
}
\value{
Called for it's side effect of writing the image to the terminal (standard
out). Returns \code{NULL} invisibly.
}
\description{
Dump an image raster to the terminal
}
\seealso{
See \code{\link{raster2tgp}} for output in Terminal Graphics Protocol.  See
\code{\link{raster2sixel}} for output in Sixel format.
}
