% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster2tgp.R
\name{raster2tgp}
\alias{raster2tgp}
\title{Dump an image raster to the terminal using the Terminal Graphics Protocol}
\usage{
raster2tgp(raster, compress = TRUE)
}
\arguments{
\item{raster}{the image 'raster' e.g. the output of \code{\link{as.raster}}.}

\item{compress}{compress the data before sending to the terminal.}
}
\value{
Called for it's side effect of writing the image to the terminal (standard
out). Returns \code{NULL} invisibly.
}
\description{
Dump an image raster to the terminal using the Terminal Graphics Protocol
}
\details{
The Terminal Graphics Protocol is not supported by many Terminal Emulators.
The most notable terminal emulator supporting the protocol is Kitty.
}
\seealso{
See \code{\link{raster2sixel}} for output in Sixel format. See
\code{\link{png2terminal}} for writing a PNG image to the terminal.
}
