% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_color_mode.R
\name{term_color_mode}
\alias{term_color_mode}
\title{Get the color mode of the terminal}
\usage{
term_color_mode()
}
\value{
Returns \code{NA} when the terminal does not support querying the color 
mode. Otherwise returns either 'dark' or 'light'.
}
\description{
Get the color mode of the terminal
}
\details{
Uses ANSI \code{SI ? 996 n} to determine the color mode. See
\url{https://contour-terminal.org/vt-extensions/color-palette-update-notifications/}.
}
\examples{
term_color_mode()

}
