% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_colors.R
\name{term_background}
\alias{term_background}
\alias{term_foreground}
\alias{term_palette}
\title{Get the colors used in the terminal}
\usage{
term_background()

term_foreground()

term_palette()
}
\value{
\code{term_background} and \code{term_foreground} will return a length 1
vector with a color. \code{term_palette} will return a vector of colors.
}
\description{
Get the colors used in the terminal
}
\details{
Getting the color palette used in the terminal is terminal specific.
Currently only the kitty terminal is supported. For other terminals default
colors are returned.
}
\examples{
term_background()

}
\seealso{
\code{\link{kitty_colors}} for the functions returning the specific colors
used in the kitty terminal.
}
