% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_dim.R
\name{term_dim}
\alias{term_dim}
\title{Get the dimensions of the terminal window in pixels}
\usage{
term_dim()
}
\value{
An integer vector with the width and height of the terminal in pixels
(\code{x_pixels} and \code{y_pixels}) and the number of text columns and rows
in the terminal window (\code{columns} and \code{rows}).

These values can be zero when the terminal does not support querying the
size. Some terminals only support querying the number of columns and rows.
Under windows the return value will always be zero as querying the size
depends on POSIX support.
}
\description{
Get the dimensions of the terminal window in pixels
}
\seealso{
\code{\link{term_width}} and \code{\link{term_height}} for only obtaining
the width and height in pixels respectively.
}
