% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_height.R
\name{term_height}
\alias{term_height}
\title{Get the height of the terminal window in pixels}
\usage{
term_height()
}
\value{
An integer with the number of pixels the terminal is high. This value can be
zero when the terminal does not support querying the size.
}
\description{
Get the height of the terminal window in pixels
}
\seealso{
\code{\link{term_width}} for the terminal width and \code{\link{term_dim}}
for all dimensions including the dimensions in rows and columns.
}
