% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_manager.R
\name{term_replot}
\alias{term_replot}
\title{Replot the current device in the terminal.}
\usage{
term_replot()
}
\value{
Called for outputting the current contents of a \code{\link{sixel}} or
\code{\link{tgp}} device into the terminal.
}
\description{
Replot the current device in the terminal.
}
