% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_width.R
\name{term_width}
\alias{term_width}
\title{Get the with of the terminal window in pixels}
\usage{
term_width()
}
\value{
An integer with the number of pixels the terminal is wide This value can be
zero when the terminal does not support querying the size. Under windows the
return value will always be zero as querying the size depends on POSIX
support.
}
\description{
Get the with of the terminal window in pixels
}
\seealso{
\code{\link{term_height}} for the terminal height and \code{\link{term_dim}}
for all dimensions including the dimensions in rows and columns.
}
