% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgp.R
\name{tgp_diacritics}
\alias{tgp_diacritics}
\title{Unicode values used by the Terminal Graphics Protocol}
\format{
A `character` vector, where each entry is a single unicode character.
  with `names()` set to the 4-digit unicode identifier.
}
\source{
[the terminal graphics protocol specification](https://sw.kovidgoyal.net/kitty/graphics-protocol/)
  and linked
  [diacritics file](https://sw.kovidgoyal.net/kitty/_downloads/f0a0de9ec8d9ff4456206db8e0814937/rowcolumn-diacritics.txt).
}
\description{
This set of combining characters ("diacritics", which are things like
accent characters and other character modifiers) to encode information for
the terminal in a way that does not render to the screen. The characters
used are very specific and are defined by the terminal graphics protocol
itself.
}
